move circular into separate package
This will prevent testing flags from appearing in normal builds, which
is caused by circular.go importing "testing" package.
